node_refs example
