+++ examples/node_refs/README.md
@@ -1,0 +1,13 @@
+# Node Refs Example
+
+This example shows two input fields which are automatically focused when hovered over.
+
+## Concepts
+
+The example uses [Refs](https://yew.rs/docs/en/concepts/components/refs/) to
+manipulate the underlying DOM element directly.
+
+## Improvements
+
+- Use CSS to improve the presentation of the example
+- Expand this to a fully fledged form with input validation etc.
